Working on InnerPattern, many data structures
InnerPattern (was IntPattern in pderiv) is slowly making progress, and
the trickiest parts so far have all been resolved. As for the data
structures... IntVLA and LongVLA have different implementations (the
first from libGDX, the other from fastutil) but probably shouldn't, and
may wind up with some code combined if possible. OMap is a "hashlist,"
sort of a variant on an insertion-ordered map that uses a list-like data
structure to store ordering (and can thus be reordered). It's taken from
SquidLib, but that in turn takes most of its code from fastutil, even
though it uses IntVLA from libGDX to handle ordering/reordering. Plus,
there's a generic Set type taken from libGDX because it has iterator
recycling.

I hope that's all we need for data structures, but I'm not yet fully
convinced...